fix: chapter state should be handled globaly
